WIP: Renamed and refactores a bunch of things
