+++ Sextant.sln
@@ -5,6 +5,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sextant", "Sextant\Sextant.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sextant.PCL", "Sextant.PCL\Sextant.PCL.csproj", "{FB81D9D6-EDBE-49AB-9591-7077E6FE5EA9}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "PCL", "PCL", "{08417F99-4EA4-4F61-84E4-978F25C70F07}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -19,5 +21,8 @@ Global
 		{FB81D9D6-EDBE-49AB-9591-7077E6FE5EA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FB81D9D6-EDBE-49AB-9591-7077E6FE5EA9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FB81D9D6-EDBE-49AB-9591-7077E6FE5EA9}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{FB81D9D6-EDBE-49AB-9591-7077E6FE5EA9} = {08417F99-4EA4-4F61-84E4-978F25C70F07}
 	EndGlobalSection
 EndGlobal
