+++ Sextant/IBaseNavigationPage.cs
@@ -1,0 +1,6 @@
+﻿namespace Sextant
+{
+    public interface IBaseNavigationPage<out TPageModel> where TPageModel : class, IBaseNavigationPageModel
+    {
+    }
+}
