--- Sextant/IBasePage.cs
@@ -1,6 +1,0 @@
-﻿namespace Sextant
-{
-    public interface IBasePage<out TPageModel> where TPageModel : class, IBasePageModel
-    {
-    }
-}
