+++ Sextant/Logger/BaseLogger.cs
