+++ Sextant/Logger/IBaseLogger.cs
