+++ Sextant/NavigationExtensions.cs
@@ -18,16 +18,16 @@ namespace Sextant
 		/// <param name="animated">If set to <c>true</c> animated.</param>
 		/// <typeparam name="TCurrentPageModel">The 1st type parameter.</typeparam>
 		/// <typeparam name="TPageModel">The 2nd type parameter.</typeparam>        
-		public static Task<bool> PushPageAsync<TCurrentPageModel, TPageModel>(this TCurrentPageModel currentPageModel, IBasePage<TPageModel> pageToPush, Action<TPageModel> executeOnPageModel = null, bool animated = true) where TCurrentPageModel : class, IBasePageModel where TPageModel : class, IBasePageModel
+		public static Task<bool> PushPageAsync<TCurrentPageModel, TPageModel>(this TCurrentPageModel currentPageModel, IBaseNavigationPage<TPageModel> pageToPush, Action<TPageModel> executeOnPageModel = null, bool animated = true) where TCurrentPageModel : class, IBaseNavigationPageModel where TPageModel : class, IBaseNavigationPageModel
 		{
-			var currentPage = SextantCore.CurrentFactory.GetPageByModel(currentPageModel);
+			var currentPage = SextantCore.Instance.GetPageByModel(currentPageModel);
 
 			if (currentPage != null)
 			{
 				if (executeOnPageModel != null)
 					pageToPush.ExecuteOnPageModel(executeOnPageModel);
 
-				return SextantCore.CurrentFactory.PushPageAsync(currentPage, pageToPush, animated);
+				return SextantCore.Instance.PushPageAsync(currentPage, pageToPush, animated);
 			}
 
 			return Task.FromResult(false);
@@ -41,18 +41,18 @@ namespace Sextant
 		/// <param name="executeOnPageModel">Execute on page model.</param>
 		/// <param name="animated">If set to <c>true</c> animated.</param>
 		/// <typeparam name="TPageModel">The 1st type parameter.</typeparam>
-		public static Task<bool> PushPageAsNewInstanceAsync<TPageModel>(this IBasePageModel currentPageModel, Action<TPageModel> executeOnPageModel = null, bool animated = true) where TPageModel : class, IBasePageModel
+		public static Task<bool> PushPageAsNewInstanceAsync<TPageModel>(this IBaseNavigationPageModel currentPageModel, Action<TPageModel> executeOnPageModel = null, bool animated = true) where TPageModel : class, IBaseNavigationPageModel
 		{
-			var currentPage = SextantCore.CurrentFactory.GetPageByModel(currentPageModel);
+			var currentPage = SextantCore.Instance.GetPageByModel(currentPageModel);
 
 			if (currentPage != null)
 			{
-				var pageToPush = SextantCore.CurrentFactory.GetPage<TPageModel>();
+				var pageToPush = SextantCore.Instance.GetPage<TPageModel>();
 
 				if (executeOnPageModel != null)
 					pageToPush.ExecuteOnPageModel(executeOnPageModel);
 
-				return SextantCore.CurrentFactory.PushPageAsync(currentPage, pageToPush, animated);
+				return SextantCore.Instance.PushPageAsync(currentPage, pageToPush, animated);
 			}
 
 			return Task.FromResult(false);
@@ -66,18 +66,18 @@ namespace Sextant
 		/// <param name="executeOnPageModel">Execute on page model.</param>
 		/// <param name="animated">If set to <c>true</c> animated.</param>
 		/// <typeparam name="TPageModel">The 1st type parameter.</typeparam>
-		public static Task<bool> PushModalPageAsNewInstanceAsync<TPageModel>(this IBasePageModel currentPageModel, Action<TPageModel> executeOnPageModel = null, bool animated = true) where TPageModel : class, IBasePageModel
+		public static Task<bool> PushModalPageAsNewInstanceAsync<TPageModel>(this IBaseNavigationPageModel currentPageModel, Action<TPageModel> executeOnPageModel = null, bool animated = true) where TPageModel : class, IBaseNavigationPageModel
 		{
-			var currentPage = SextantCore.CurrentFactory.GetPageByModel(currentPageModel);
+			var currentPage = SextantCore.Instance.GetPageByModel(currentPageModel);
 
 			if (currentPage != null)
 			{
-				var pageToPush = SextantCore.CurrentFactory.GetPage<TPageModel>();
+				var pageToPush = SextantCore.Instance.GetPage<TPageModel>();
 
 				if (executeOnPageModel != null)
 					pageToPush.ExecuteOnPageModel(executeOnPageModel);
 
-				return SextantCore.CurrentFactory.PushModalPageAsync(currentPage, pageToPush, animated);
+				return SextantCore.Instance.PushModalPageAsync(currentPage, pageToPush, animated);
 			}
 
 			return Task.FromResult(false);
@@ -93,16 +93,16 @@ namespace Sextant
 		/// <param name="animated">If set to <c>true</c> animated.</param>
 		/// <typeparam name="TCurrentPageModel">The 1st type parameter.</typeparam>
 		/// <typeparam name="TPageModel">The 2nd type parameter.</typeparam>
-		public static Task<bool> PushModalPageAsync<TCurrentPageModel, TPageModel>(this TCurrentPageModel currentPageModel, IBasePage<TPageModel> pageToPush, Action<TPageModel> executeOnPageModel = null, bool animated = true) where TCurrentPageModel : class, IBasePageModel where TPageModel : class, IBasePageModel
+		public static Task<bool> PushModalPageAsync<TCurrentPageModel, TPageModel>(this TCurrentPageModel currentPageModel, IBaseNavigationPage<TPageModel> pageToPush, Action<TPageModel> executeOnPageModel = null, bool animated = true) where TCurrentPageModel : class, IBaseNavigationPageModel where TPageModel : class, IBaseNavigationPageModel
 		{
-			var currentPage = SextantCore.CurrentFactory.GetPageByModel(currentPageModel);
+			var currentPage = SextantCore.Instance.GetPageByModel(currentPageModel);
 
 			if (currentPage != null)
 			{
 				if (executeOnPageModel != null)
 					pageToPush.ExecuteOnPageModel(executeOnPageModel);
 
-				return SextantCore.CurrentFactory.PushModalPageAsync(currentPage, pageToPush, animated);
+				return SextantCore.Instance.PushModalPageAsync(currentPage, pageToPush, animated);
 			}
 
 			return Task.FromResult(false);
@@ -115,12 +115,12 @@ namespace Sextant
 		/// <param name="pageModel">Page model.</param>
 		/// <param name="animated">If set to <c>true</c> animated.</param>
 		/// <typeparam name="TPageModel">The 1st type parameter.</typeparam>
-		public static Task<bool> PopPageAsync<TPageModel>(this TPageModel pageModel, bool animated = true) where TPageModel : class, IBasePageModel
+		public static Task<bool> PopPageAsync<TPageModel>(this TPageModel pageModel, bool animated = true) where TPageModel : class, IBaseNavigationPageModel
 		{
-			var page = SextantCore.CurrentFactory.GetPageByModel(pageModel);
+			var page = SextantCore.Instance.GetPageByModel(pageModel);
 
 			if (page != null)
-				return SextantCore.CurrentFactory.PopPageAsync(page, animated);
+				return SextantCore.Instance.PopPageAsync(page, animated);
 
 			return Task.FromResult(false);
 		}
@@ -132,12 +132,12 @@ namespace Sextant
 		/// <param name="pageModel">Page model.</param>
 		/// <param name="animated">If set to <c>true</c> animated.</param>
 		/// <typeparam name="TPageModel">The 1st type parameter.</typeparam>
-		public static Task<bool> PopModalPageAsync<TPageModel>(this TPageModel pageModel, bool animated = true) where TPageModel : class, IBasePageModel
+		public static Task<bool> PopModalPageAsync<TPageModel>(this TPageModel pageModel, bool animated = true) where TPageModel : class, IBaseNavigationPageModel
 		{
-			var page = SextantCore.CurrentFactory.GetPageByModel(pageModel);
+			var page = SextantCore.Instance.GetPageByModel(pageModel);
 
 			if (page != null)
-				return SextantCore.CurrentFactory.PopModalPageAsync(page, animated);
+				return SextantCore.Instance.PopModalPageAsync(page, animated);
 
 			return Task.FromResult(false);
 		}
@@ -150,12 +150,12 @@ namespace Sextant
 		/// <param name="pageToRemove">Page to remove.</param>
 		/// <typeparam name="TCurrentPageModel">The 1st type parameter.</typeparam>
 		/// <typeparam name="TPageModel">The 2nd type parameter.</typeparam>
-		public static Task<bool> RemovePageAsync<TCurrentPageModel, TPageModel>(this TCurrentPageModel currentPageModel, IBasePage<TPageModel> pageToRemove) where TCurrentPageModel : class, IBasePageModel where TPageModel : class, IBasePageModel
+		public static Task<bool> RemovePageAsync<TCurrentPageModel, TPageModel>(this TCurrentPageModel currentPageModel, IBaseNavigationPage<TPageModel> pageToRemove) where TCurrentPageModel : class, IBaseNavigationPageModel where TPageModel : class, IBaseNavigationPageModel
 		{
-			var currentPage = SextantCore.CurrentFactory.GetPageByModel(currentPageModel);
+			var currentPage = SextantCore.Instance.GetPageByModel(currentPageModel);
 
 			if (currentPage != null)
-				return SextantCore.CurrentFactory.RemovePageAsync(currentPage, pageToRemove);
+				return SextantCore.Instance.RemovePageAsync(currentPage, pageToRemove);
 
 			return Task.FromResult(false);
 		}
@@ -167,12 +167,12 @@ namespace Sextant
 		/// <param name="currentPageModel">Current page model.</param>
 		/// <param name="animated">If set to <c>true</c> animated.</param>
 		/// <typeparam name="TCurrentPageModel">The 1st type parameter.</typeparam>
-		public static Task<bool> PopPagesToRootAsync<TCurrentPageModel>(this TCurrentPageModel currentPageModel, bool animated = true) where TCurrentPageModel : class, IBasePageModel
+		public static Task<bool> PopPagesToRootAsync<TCurrentPageModel>(this TCurrentPageModel currentPageModel, bool animated = true) where TCurrentPageModel : class, IBaseNavigationPageModel
 		{
-			var currentPage = SextantCore.CurrentFactory.GetPageByModel(currentPageModel);
+			var currentPage = SextantCore.Instance.GetPageByModel(currentPageModel);
 
 			if (currentPage != null)
-				return SextantCore.CurrentFactory.PopPagesToRootAsync(currentPage, animated);
+				return SextantCore.Instance.PopPagesToRootAsync(currentPage, animated);
 
 			return Task.FromResult(false);
 		}
@@ -182,42 +182,42 @@ namespace Sextant
 		/// Sets the new root and resets based on PageModel
 		/// </summary>
 		/// <param name="currentPageModel">Current page model.</param>
-		public static Task<bool> SetNewRootAndResetAsync<TNewRootPageModel>(this IBasePageModel currentPageModel) where TNewRootPageModel : class, IBasePageModel
+		public static Task<bool> SetNewRootAndResetAsync<TNewRootPageModel>(this IBaseNavigationPageModel currentPageModel) where TNewRootPageModel : class, IBaseNavigationPageModel
 		{
-			return SextantCore.CurrentFactory.SetNewRootAndResetAsync<TNewRootPageModel>();
+			return SextantCore.Instance.SetNewRootAndResetAsync<TNewRootPageModel>();
 		}
 
-		public static Task<bool> PushPageAsNewInstanceAsync<TPageNavigationModel, TViewModel>(this IBasePageModel currenTViewModel, Action<TViewModel> executeOnPageModel = null, bool animated = true) where TViewModel : class, IBasePageModel where TPageNavigationModel : class, IBasePageModel
+		public static Task<bool> PushPageAsNewInstanceAsync<TPageNavigationModel, TViewModel>(this IBaseNavigationPageModel currenTViewModel, Action<TViewModel> executeOnPageModel = null, bool animated = true) where TViewModel : class, IBaseNavigationPageModel where TPageNavigationModel : class, IBaseNavigationPageModel
 		{
-			var currentPage = SextantCore.CurrentFactory.GetPageByModel(currenTViewModel);
+			var currentPage = SextantCore.Instance.GetPageByModel(currenTViewModel);
 
 			if (currentPage != null)
 			{
-				var pageToPush = SextantCore.CurrentFactory.GetPage<TViewModel>();
-				var navigationPageToPush = SextantCore.CurrentFactory.GetPage<TPageNavigationModel>();
+				var pageToPush = SextantCore.Instance.GetPage<TViewModel>();
+				var navigationPageToPush = SextantCore.Instance.GetPage<TPageNavigationModel>();
 
 				if (executeOnPageModel != null)
 					pageToPush.ExecuteOnPageModel(executeOnPageModel);
 
-				return SextantCore.CurrentFactory.PushPageAsync(currentPage, navigationPageToPush, animated);
+				return SextantCore.Instance.PushPageAsync(currentPage, navigationPageToPush, animated);
 			}
 
 			return Task.FromResult(false);
 		}
 
-		public static Task<bool> PushModalPageAsNewInstanceAsync<TPageNavigationModel, TViewModel>(this IBasePageModel currenTViewModel, Action<TViewModel> executeOnPageModel = null, bool animated = true) where TViewModel : class, IBasePageModel where TPageNavigationModel : class, IBasePageModel
+		public static Task<bool> PushModalPageAsNewInstanceAsync<TPageNavigationModel, TViewModel>(this IBaseNavigationPageModel currenTViewModel, Action<TViewModel> executeOnPageModel = null, bool animated = true) where TViewModel : class, IBaseNavigationPageModel where TPageNavigationModel : class, IBaseNavigationPageModel
 		{
-			var currentPage = SextantCore.CurrentFactory.GetPageByModel(currenTViewModel);
+			var currentPage = SextantCore.Instance.GetPageByModel(currenTViewModel);
 
 			if (currentPage != null)
 			{
-				var pageToPush = SextantCore.CurrentFactory.GetPage<TViewModel>();
-				var navigationPageToPush = SextantCore.CurrentFactory.GetPage<TPageNavigationModel>();
+				var pageToPush = SextantCore.Instance.GetPage<TViewModel>();
+				var navigationPageToPush = SextantCore.Instance.GetPage<TPageNavigationModel>();
 
 				if (executeOnPageModel != null)
 					pageToPush.ExecuteOnPageModel(executeOnPageModel);
 
-				return SextantCore.CurrentFactory.PushModalPageAsync(currentPage, navigationPageToPush, animated);
+				return SextantCore.Instance.PushModalPageAsync(currentPage, navigationPageToPush, animated);
 			}
 
 			return Task.FromResult(false);
