+++ Sextant/SextantCore.cs
@@ -1,43 +1,28 @@
 ﻿using System;
 namespace Sextant
 {
-	/// <summary>
-	/// Page factory.
-	/// </summary>
-	public static class SextantCore
-	{
-		static IBaseFactory current;
+    public static class SextantCore
+    {
+        public static IBaseLogger Logger { get; set; }
 
-		/// <summary>
-		/// Gets or sets the logger.
-		/// </summary>
-		/// <value>The logger.</value>
-		public static IBaseLogger Logger { get; set; }
-        
-		/// <summary>
-		/// Gets the Factory instance.
-		/// </summary>
-		/// <value>The factory.</value>
-		public static IBaseFactory CurrentFactory
-		{
-			get
-			{
-				if (current == null)
-				{
-					throw new NullReferenceException("CurrentFactory is null. Please initialize it with SetCurrentFactory method");
-				}
+        static ISextantNavigationService current;
 
-				return current;
-			}
-		}
+        public static ISextantNavigationService Instance
+        {
+            get
+            {
+                if (current == null)
+                {
+                    throw new NullReferenceException("CurrentFactory is null. Please initialize it with SetCurrentFactory method");
+                }
 
-		/// <summary>
-		/// Initializes Factory.
-		/// </summary>
-		/// <param name="factory">Factory.</param>
-		public static void SetCurrentFactory(IBaseFactory factory)
-		{
-			current = factory;
-		}
-	}
+                return current;
+            }
+        }
+
+        public static void SetCurrentFactory(ISextantNavigationService factory)
+        {
+            current = factory;
+        }
+    }
 }
